Update deadline after sending the hard timeout signal
Otherwise it's very likely the monitor loop will just issue the
same signal again an again until the process die.

The signal being SIGKILL it's not that big of a deal, but might
as well not spam the kernel with non-sense.

Also in some scenarios it may be useful to replace the signal with
SIGABRT to get a core dump to debug why the hard timeout was invoked
in the first place. In such situation the spamming is harmful.